+++ frontend/cashcompass_hook/lib/src/connector/connector.dart
