+++ frontend/cashcompass_hook/lib/src/dtos/data_class.dart
@@ -1,4 +1,4 @@
-import 'package:cashcompass_hook/src/connector/connectorImpl.dart';
+import 'package:cashcompass_hook/src/connector/connector.dart';
 import 'package:cashcompass_hook/src/dtos/base_dto.dart';
 
 abstract class DataClass<T extends BaseDTO> {
